move mount prepartions to own class
